Add the intercom "ringing" signal
The red intercom wire is now connected to an ESP32 input through a
voltage divider to detect the exact moment the caller presses the
intercom button, reflect this in the config.  In the future we may be
able to even distinguish the one ring signal (staircase door) from the
other ring signal (street level gate) based on the waveform.