intercom: Don't play call-end sound when silenced
Respect the silenced state.  In theory when silenced is on, the ground
of the speaker is not connected and it shouldn't make any sound but
there may be some current flowing through somewhere and the sound does
actually come out at a low volume.

Where is the current escaping through?
 * Red is the speaker positive.
 * Green is disconnected externally when call not active and
   additionally is disconnected internally when phone is hung up.
 * White is disconnected externally when call not active and
   additionally is disconnected internally when phone is hung up.
 * Yellow should be disconnected externally when call not active
   and at 12V when active, and disconnected internally until the door
   open button is pressed.
 * Blue is disconnected by the mosfet in the intercom node, maybe it's
   not opening the circuit completely? why?  Mosfet gate should only be
   connected to GPIO12 and a pulldown.