bedroom1: Add virtual switch to disable shutters control
Add virtual switch to disable shutters control with physical switch (kids).

Refactor the night light code a little to try to fix an issue with it
but that on its own is not fixing it, just making it possibly easier to
debug.  So far seems to be an esphome issue?